requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
fastapi==0.110.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
minio==7.2.5
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
setuptools==69.2.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
zope.event==5.0
zope.interface==6.2
